backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.0
celery>=5.3.0
redis>=5.0.0
django-filter>=23.0.0
Pillow>=10.0.0
python-decouple>=3.8
